fix(iFrame): Make resizeLargeVideo only available on web
