+++ react/features/large-video/actions.native.js
@@ -1,0 +1,3 @@
+// @flow
+
+export * from './actions.any';
